moved game component
