+++ src/routes/AppRoutes.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Home from "../components/common/Home";
 import Navbar from "../components/common/Navbar";
-import MemoryBoxGame from "../components/questions/MemoryBoxGame";
+import MemoryBoxGame from "../components/questions/MemoryBox/MemoryBoxGame";
 import PageNotFound from "../components/common/PageNotFound";
 import LoginPage from "../components/common/LoginPage";
 import ToastNotification from "../components/questions/ToastNotification/ToastNotificationContainer";
